refactor Faze.Rendering.Video
