--- src/libraries/Faze.Rendering.Video/Faze.Rendering.Video/Extensions/VideoFFMPEGSettings.cs
@@ -1,12 +1,0 @@
-﻿namespace Faze.Rendering.Video.Extensions
-{
-    public class VideoFFMPEGSettings
-    {
-        public VideoFFMPEGSettings(int fps)
-        {
-            Fps = fps;
-        }
-
-        public int Fps { get; }
-    }
-}
