+++ src/libraries/Faze.Rendering.Video/Faze.Rendering.Video/FFMPEGVCodec.cs
@@ -1,0 +1,8 @@
+﻿namespace Faze.Rendering.Video.Extensions
+{
+    public enum FFMPEGVCodec 
+    {
+        Png,
+        Mjpeg,
+    }
+}
